Made hash functions able to deal with keys with negative numbers.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@14066 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/elf.c
@@ -69,7 +69,7 @@ image_hash(void *_image, const void *_key, uint32 range)
 	if (image != NULL)
 		return image->id % range;
 
-	return id % range;
+	return (uint32)id % range;
 }
